feat(DOM): Ships placed into player 1s gameboard in the DOM.
Player2s board is accessible. Need to change player 2s board to be hidden, alternate turns with
player 1 if both human.
If AI, then player2 is to always be hidden and solely used for attacking on player 1 turn.

Bugs:
player.type is returning undefined, needs to be "human" or "ai" to determine how to handle future interactions.
Ships are being placed out bounds if the first coord is in bounds, but the others are not.
For example, ship.length = 4; startCoords = [9,9] => the remaining coords would go out of bounds.